Script for switching from wpa_supplicant to iwd
I was trying to investigate for several hours why after router restarts,
NetworkManager is unable to pick connection back up.
It dies with this reconnection log:
no secrets: No agents were available for this request.
Even though auth prompt had password correctly filled in.
It most likely has something to ho with keyring and nm-applet
and its interoperability, but I was unable to find out how to fix it.
I found one post when searching to switch to iwd backend.
I tried that and it magically works, so I have written this script
to make it so if I were to reinstall my OS and I needed to do it again.